collapse button refactorized
